pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "consult7"
version = "1.3.0"
description = "MCP server for consulting large context window models to analyze extensive file collections"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Stefan Szeider", email = "stefan@szeider.net"}
]
keywords = ["mcp", "llm", "code-analysis", "large-context"]
dependencies = [
    "mcp>=1.9.4",
    "httpx>=0.28.1",
    "google-genai>=1.19.0",
    "openai>=1.88.0",
    "PyYAML>=6.0",
    "pydantic>=2.0.0",
    "platformdirs>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/szeider/consult7"
Repository = "https://github.com/szeider/consult7"
Issues = "https://github.com/szeider/consult7/issues"

[project.scripts]
consult7 = "consult7.server:run"

[tool.hatch.build.targets.wheel]
packages = ["src/consult7"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
    "anthropic>=0.52.0",  # For MCP testing support
]

[tool.ruff]
lint.select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "T201",   # flake8-print: print found
]
lint.ignore = []

[tool.ruff.lint.per-file-ignores]
# Allow print statements in test files
"tests/**/*.py" = ["T201"]
# Allow print statements in server.py for CLI output and test mode
"src/consult7/server.py" = ["T201"]
